move to src/plugins/remark-extract-image-class
