+++ src/plugins/remark-extract-image-class.js
